Move helper methods to TableMethods support module
Now spec_helper only contains configuration
